Extract db components to db layer
Extract flights table logic from service
