+++ src/api/db/models/flights_model.py
@@ -1,7 +1,7 @@
 import sqlalchemy
 from sqlalchemy.orm import relationship
 
-from api.database import db
+from api.db.database import db
 
 
 class Flights(db.Model):
