+++ src/api/db/models/users_model.py
@@ -1,6 +1,6 @@
 import sqlalchemy
 from sqlalchemy.orm import relationship
-from api.database import db
+from api.db.database import db
 
 
 class Users(db.Model):
